// indexers/enterprise/src/types/enterprise_facade.ts
export type Uint128 = string
export type Addr = string
// ...
export type Timestamp = Uint64
// ...
export type Expiration =
  | {
      at_height: number
    }
  | {
      at_time: Timestamp
    }
  | {
      never: {}
    }
export type ProposalAction =
  | {
      update_metadata: UpdateMetadataMsg
    }
  | {
      update_gov_config: UpdateGovConfigMsg
    }
  | {
      update_council: UpdateCouncilMsg
    }
  | {
      update_asset_whitelist: UpdateAssetWhitelistProposalActionMsg
    }
  | {
      update_nft_whitelist: UpdateNftWhitelistProposalActionMsg
    }
  | {
      request_funding_from_dao: RequestFundingFromDaoMsg
    }
  | {
      upgrade_dao: UpgradeDaoMsg
    }
  | {
      execute_msgs: ExecuteMsgsMsg
    }
  | {
      modify_multisig_membership: ModifyMultisigMembershipMsg
    }
  | {
      distribute_funds: DistributeFundsMsg
    }
  | {
      update_minimum_weight_for_rewards: UpdateMinimumWeightForRewardsMsg
    }
  | {
      add_attestation: AddAttestationMsg
    }
  | {
      remove_attestation: {}
    }
  | {
      deploy_cross_chain_treasury: DeployCrossChainTreasuryMsg
    }
// ...
export type AssetInfoBaseFor_String =
  | {
      native: string
    }
  | {
      cw20: string
    }
  | {
      /**
       * @minItems 2
       * @maxItems 2
       */
      cw1155: [string, string]
    }
export type Binary = string
export type ProposalType = 'general' | 'council'
export type ProposalStatus = 'in_progress' | 'passed' | 'rejected' | 'executed'
export interface ProposalResponse {
  proposal: Proposal
  proposal_status: ProposalStatus
  /**
   * Total vote-count (value) for each outcome (key).
   */
  results: [number, Uint128][]
  total_votes_available: Uint128
}
export interface Proposal {
  description: string
  expires: Expiration
  id: number
  proposal_actions: ProposalAction[]
  proposal_type: ProposalType
  proposer: Addr
  started_at: Timestamp
  status: ProposalStatus
  title: string
}
// ...
export type QueryMsg =
  | {
      dao_info: {
        contract: Addr
      }
    }
  | {
      member_info: {
        contract: Addr
        msg: QueryMemberInfoMsg
      }
    }
  | {
      list_multisig_members: {
        contract: Addr
        msg: ListMultisigMembersMsg
      }
    }
  | {
      asset_whitelist: {
        contract: Addr
        params: AssetWhitelistParams
      }
    }
  | {
      nft_whitelist: {
        contract: Addr
        params: NftWhitelistParams
      }
    }
  | {
      proposal: {
        contract: Addr
        params: ProposalParams
      }
    }
  | {
      proposals: {
        contract: Addr
        params: ProposalsParams
      }
    }
  | {
      proposal_status: {
        contract: Addr
        params: ProposalStatusParams
      }
    }
  | {
      member_vote: {
        contract: Addr
        params: MemberVoteParams
      }
    }
  | {
      proposal_votes: {
        contract: Addr
        params: ProposalVotesParams
      }
    }
  | {
      user_stake: {
        contract: Addr
        params: UserStakeParams
      }
    }
  | {
      total_staked_amount: {
        contract: Addr
      }
    }
  | {
      staked_nfts: {
        contract: Addr
        params: StakedNftsParams
      }
    }
  | {
      claims: {
        contract: Addr
        params: ClaimsParams
      }
    }
  | {
      releasable_claims: {
        contract: Addr
        params: ClaimsParams
      }
    }
  | {
      create_proposal_adapted: {
        contract: Addr
        params: CreateProposalMsg
      }
    }
  | {
      create_proposal_with_denom_deposit_adapted: {
        contract: Addr
        params: CreateProposalWithDenomDepositMsg
      }
    }
  | {
      create_proposal_with_token_deposit_adapted: {
        contract: Addr
        params: CreateProposalWithTokenDepositMsg
      }
    }
  | {
      create_proposal_with_nft_deposit_adapted: {
        contract: Addr
        params: CreateProposalWithNftDepositMsg
      }
    }
  | {
      create_council_proposal_adapted: {
        contract: Addr
        params: CreateProposalMsg
      }
    }
  | {
      cast_vote_adapted: {
        contract: Addr
        params: CastVoteMsg
      }
    }
  | {
      cast_council_vote_adapted: {
        contract: Addr
        params: CastVoteMsg
      }
    }
  | {
      stake_adapted: {
        contract: Addr
        params: StakeMsg
      }
    }
  | {
      unstake_adapted: {
        contract: Addr
        params: UnstakeMsg
      }
    }
  | {
      claim_adapted: {
        contract: Addr
      }
    }
export type ProposalStatusFilter = 'in_progress' | 'passed' | 'rejected'
export type VoteOutcome = 'yes' | 'no' | 'abstain' | 'veto'
export type StakeMsg =
  | {
      cw20: StakeCw20Msg
    }
  | {
      cw721: StakeCw721Msg
    }
  | {
      denom: StakeDenomMsg
    }
export type UnstakeMsg =
  | {
      cw20: UnstakeCw20Msg
    }
  | {
      cw721: UnstakeCw721Msg
    }
export interface QueryMemberInfoMsg {
  member_address: string
}
export interface ListMultisigMembersMsg {
  limit?: number | null
  start_after?: string | null
}
export interface AssetWhitelistParams {
  limit?: number | null
  start_after?: AssetInfoBaseFor_String | null
}
export interface NftWhitelistParams {
  limit?: number | null
  start_after?: string | null
}
export interface ProposalParams {
  proposal_id: number
}
export interface ProposalsParams {
  /**
   * Optional proposal status to filter for.
   */
  filter?: ProposalStatusFilter | null
  limit?: number | null
  start_after?: number | null
}
export interface ProposalStatusParams {
  proposal_id: number
}
export interface MemberVoteParams {
  member: string
  proposal_id: number
}
export interface ProposalVotesParams {
  limit?: number | null
  proposal_id: number
  /**
   * Optional pagination data, will return votes after the given voter address
   */
  start_after?: string | null
}
export interface UserStakeParams {
  limit?: number | null
  start_after?: string | null
  user: string
}
export interface StakedNftsParams {
  limit?: number | null
  start_after?: string | null
}
export interface ClaimsParams {
  owner: string
}
export interface CreateProposalMsg {
  /**
   * Optional description text of the proposal
   */
  description?: string | null
  /**
   * Actions to be executed, in order, if the proposal passes
   */
  proposal_actions: ProposalAction[]
  /**
   * Title of the proposal
   */
  title: string
}
export interface CreateProposalWithDenomDepositMsg {
  create_proposal_msg: CreateProposalMsg
  deposit_amount: Uint128
}
export interface CreateProposalWithTokenDepositMsg {
  create_proposal_msg: CreateProposalMsg
  deposit_amount: Uint128
}
export interface CreateProposalWithNftDepositMsg {
  create_proposal_msg: CreateProposalMsg
  /**
   * Tokens that the user wants to deposit to create the proposal. These tokens are expected to be owned by the user or approved for them, otherwise this fails. governance-controller expects to have an approval for those tokens.
   */
  deposit_tokens: string[]
}
export interface CastVoteMsg {
  outcome: VoteOutcome
  proposal_id: number
}
export interface StakeCw20Msg {
  amount: Uint128
  user: string
}
export interface StakeCw721Msg {
  tokens: string[]
  user: string
}
export interface StakeDenomMsg {
  amount: Uint128
}
export interface UnstakeCw20Msg {
  amount: Uint128
}
export interface UnstakeCw721Msg {
  tokens: string[]
}